# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Registro de Pontos",
        "description": "API para gerenciar pontos dos funcionários",
        "contact": {
            "email": "diegohsananias@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/users": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Cadastrar um novo usuário",
                "description": "Cadastrar um novo usuário (admin, funcionário ou cliente).",
                "operationId": "createUser",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Carlos Silva"
                                    },
                                    "cpf": {
                                        "type": "string",
                                        "maxLength": 14,
                                        "example": "123.456.789-09"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "carlos.silva@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "minLength": 8,
                                        "example": "senha123"
                                    },
                                    "role": {
                                        "type": "string",
                                        "enum": [
                                            "admin",
                                            "user",
                                            "employee"
                                        ],
                                        "example": "employee"
                                    },
                                    "birthDate": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1990-05-20"
                                    },
                                    "cep": {
                                        "type": "string",
                                        "maxLength": 9,
                                        "example": "01001-000"
                                    },
                                    "address": {
                                        "properties": {
                                            "street": {
                                                "type": "string",
                                                "maxLength": 255,
                                                "example": "Rua Exemplo"
                                            },
                                            "city": {
                                                "type": "string",
                                                "maxLength": 255,
                                                "example": "Cidade Exemplo"
                                            },
                                            "state": {
                                                "type": "string",
                                                "maxLength": 255,
                                                "example": "Estado Exemplo"
                                            },
                                            "number": {
                                                "type": "integer",
                                                "example": 123
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuário criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Carlos Silva"
                                        },
                                        "cpf": {
                                            "type": "string",
                                            "example": "123.456.789-09"
                                        },
                                        "email": {
                                            "type": "string",
                                            "example": "carlos.silva@example.com"
                                        },
                                        "role": {
                                            "type": "string",
                                            "example": "employee"
                                        },
                                        "birthDate": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "1990-05-20"
                                        },
                                        "address": {
                                            "properties": {
                                                "cep": {
                                                    "type": "string",
                                                    "example": "01001-000"
                                                },
                                                "street": {
                                                    "type": "string",
                                                    "example": "Rua Exemplo"
                                                },
                                                "city": {
                                                    "type": "string",
                                                    "example": "Cidade Exemplo"
                                                },
                                                "state": {
                                                    "type": "string",
                                                    "example": "Estado Exemplo"
                                                },
                                                "number": {
                                                    "type": "integer",
                                                    "example": 123
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Dados inválidos"
                    },
                    "409": {
                        "description": "CPF ou e-mail já está em uso"
                    }
                }
            }
        },
        "/change-password": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Alterar senha do usuário autenticado",
                "description": "Alterar a senha do usuário autenticado. O usuário deve informar a senha atual e a nova senha.",
                "operationId": "changePassword",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "current_password": {
                                        "type": "string",
                                        "format": "password",
                                        "minLength": 8,
                                        "example": "senha123"
                                    },
                                    "new_password": {
                                        "type": "string",
                                        "format": "password",
                                        "minLength": 8,
                                        "example": "novaSenha123"
                                    },
                                    "new_password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "minLength": 8,
                                        "example": "novaSenha123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Senha alterada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Senha alterada com sucesso"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Dados inválidos"
                    },
                    "401": {
                        "description": "Senha atual incorreta"
                    }
                }
            }
        }
    },
...
}